LCM: work on implementing exact mass for composite tet elements.
Adding PHAL_ScatterCompositeTetMassRank1 Kokkos policy, which scatters
the exact local mass (not yet scaled by the element mass, but will be)
times n_coeff into the global Jacobian.

This has not been tested and more stuff needs to be done to hook everything up into the code,
so this is WIP.

Also need to figure out how residual is affected by mass matrix which isn't obvious to
me right now.