LCM: putting in logic to turn off acceleration terms when using composite tets + dynamics, to use exact mass supplied within Albany rather than compute the mass using AD.
Logic is similar to what is done in Peridigm (see e.g., albanyIsCreatingMassMatrix-based
logic in PeridigmForce_Def.hpp).

Code is not ready to be tested yet, so commented out for now.

I keep forgetting to reference the relevant issue, which is #230.
The entire branch lcm-comp-tet-exact-mass is devoted to this issue.