PHAL Removed option of scattering a scalar nodal parameter only once.
It is wrong if there's more than one workset. The correct way to update
the 'do_scatter' variable would be in the postEvaluate method, after all
worksets have been processed. However, this would require to call the
postEvaluate method on the field managers, from Albany::Application.
But Albany::Application stores field managers depending on the physical
block, which depends on the ws index. In short, it would require a few
changes to Albany::Application (in addition to the evaluator), so for now
we simply remove the feature.